refactor(webui): tighten LLM config review fixes
